Extract delegate class
